# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Chekdee API
  description: >
    Employee attendance API for Chekdee โดย turnPRO.
    This spec is the shared contract between the Go backend and the
    generated frontend client — see design.md and CONTRIBUTING.md for
    project-wide conventions.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /auth/line/authorize:
    get:
      operationId: lineAuthorize
      summary: >
        Begin a LINE login. Mints a CSRF state and an id_token nonce, binds
        them to short-lived httpOnly cookies, and returns the LINE
        authorization URL to redirect the user to. The state is echoed back as
        a query parameter on the callback and must be sent to /auth/line/login.
      parameters:
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Authorization URL issued (sets state and nonce cookies)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineAuthorizeResponse"
        "400":
          description: Missing redirect_uri
  /auth/line/login:
    post:
      operationId: lineLogin
      summary: Exchange a LINE Login authorization code for a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LineLoginRequest"
      responses:
        "200":
          description: >
            Login succeeded. Sets an httpOnly refresh-token cookie
            (scoped to /auth) in addition to the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: LINE code/profile exchange failed
        "403":
          description: Account is deactivated
  /auth/login:
    post:
      operationId: passwordLogin
      summary: >
        Username/password login for the system_owner role — the only
        role that doesn't use LINE login. Accounts are bootstrapped
        out-of-band (see backend/cmd/seedowner), there is no signup
        endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordLoginRequest"
      responses:
        "200":
          description: >
            Login succeeded. Sets an httpOnly refresh-token cookie
            (scoped to /auth) in addition to the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing username or password
        "401":
          description: Invalid username or password
        "403":
          description: Account is deactivated
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate the refresh-token cookie for a new access token
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [access_token, expires_at]
                properties:
                  access_token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
        "401":
          description: Missing, invalid, or expired refresh token
...
  /auth/register:
    post:
      operationId: completeRegistration
      summary: Complete first-time onboarding for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteRegistrationRequest"
      responses:
        "200":
          description: Registration completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Missing required fields
        "401":
          description: Missing or invalid access token
...
  /attendance/check-in:
    post:
      operationId: checkIn
      summary: >
        Verify a scanned QR payload (signature, expiry, single-use) and
        record the check-in, computing the no-grace-period late/absent
        status against the employee's schedule for today.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Missing qr_token or idempotency_key
        "401":
          description: Missing/invalid access token, or the qr_token is invalid/expired/device inactive
        "409":
          description: QR code already used, or already checked in today
  /attendance/check-out:
    post:
      operationId: checkOut
      summary: >
        Plain in-app checkout — no QR, no camera step, per the flow update
        that replaced the original checkout design.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutRequest"
      responses:
        "200":
          description: Check-out recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Missing idempotency_key
        "401":
          description: Missing or invalid access token
        "409":
          description: Not checked in today, or already checked out today
...
components:
...
  schemas:
...
    LineAuthorizeResponse:
      type: object
      required: [authorize_url]
      properties:
        authorize_url:
          type: string
    LineLoginRequest:
      type: object
      required: [code, redirect_uri, state]
      properties:
        code:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
          description: Must equal the state returned by /auth/line/authorize.
    PasswordLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    CompleteRegistrationRequest:
      type: object
      required: [first_name, last_name, student_gen]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        student_gen:
          type: string
        student_id:
          type: string
          description: Optional -- adding it as required would break registration for anyone already mid-flow.
        phone_number:
          type: string
          description: Optional, same reasoning as student_id.
    User:
      type: object
      required: [id, role, is_registered]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [system_owner, admin, supervisor, employee]
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        student_gen:
          type: string
          nullable: true
        student_id:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
        display_name:
          type: string
          nullable: true
        picture_url:
          type: string
          nullable: true
        is_registered:
          type: boolean
    AuthResponse:
      type: object
      required: [access_token, expires_at, user]
      properties:
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    CheckInRequest:
      type: object
      required: [qr_token, idempotency_key]
      properties:
        qr_token:
          type: string
        idempotency_key:
          type: string
          description: Client-generated, stable across retries of the same logical attempt.
    CheckOutRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, employee_id, work_date, status, auto_closed]
      properties:
        id:
          type: string
        employee_id:
          type: string
        work_date:
          type: string
          format: date
        check_in_at:
          type: string
          nullable: true
          example: "09:03:00"
        check_out_at:
          type: string
          nullable: true
          example: "17:05:00"
        status:
          type: string
          enum: [pending, present, late, absent]
        auto_closed:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT